Geometry Overlaps
Small geometry overlaps found for DESCANT and GRIFFIN. The green, yellow
and blue detectors had small overlaps, this is probably due to taking
the pure mathematical model of DESCANT, and then Saint Gobain rounding
the results to generate the CAD files. We need to squeeze the green and
yellow detectors along the y direction, and we need to squeeze the blue
detectors on one x edge of the blue detector. The overlaps are on the
order of 10 um, very small. See code.

The "end_plates" in GRIFFIN did not have rounded corners, which
conflicted with the rounded corners of the can. A couple of quick cuts
solved this issue. And the extension suppressors had small overlaps with
the suppressor shell. The cuts were modified to fix the overlap. Again,
very very small overlaps.