Small Changes
Made the additional  variable delrin sphere around the 8pi vacuum
chamber its own unique logical volume so that it can be used in
conjunction with the original 8pi vacuum chamber. The GRIFFIN LN2 tank
was originally just a solid cylinder of Al, I've now made it an Al shell
filled with LN2.  And finally added positrons to the efficiency particle
generator.